Add DangerColorEditor, separate from edit page
