+++ src/components/layout/BaseColors/index.ts
@@ -1,4 +1,5 @@
 export * from './ColorListItem';
+export * from './DangerColorEditor';
 export * from './DebouncedColorPicker';
 export * from './NeutralColorEditor';
 export * from './PalettePreview';
